+++ src/Components/LandingPage/LandingPage.css
@@ -1,21 +1,23 @@
-.results-window {
-  border: 2px solid sienna;
+.provider-box {
+  border: 2px solid indigo;
   width: 80vw;
   display: flex;
   flex-direction: column;
-  align-items: center;
 }
 
-.filter-box {
+.provider-background {
+  border: 1px solid orange;
   display: flex;
-  flex-direction: column;
-  align-items: center;
+  justify-content: space-around;
 }
 
-.results-cards {
-  border: 2px solid salmon;
-  width: 80vw;
+.provider-information {
+  border: 1px solid blue;
+}
+
+.provider-tags {
+  border: 1px solid red;
   display: flex;
-  flex-wrap: wrap;
+  align-items: center;
   justify-content: space-evenly;
 }
