+++ src/Components/TherapistCard/TherapistCard.js
@@ -1,15 +1,15 @@
-import React, {Component} from "react";
+import React from "react";
 import "./TherapistCard.css"
 
-const TherapistCard = ({ id, name }) => {
+const TherapistCard = ({ id, name, labels, imageUrl }) => {
 
     return (
       <div className="therapist-card">
-        <h2>Picture will go here </h2>
+        <img src={imageUrl}/>
         <p>Name: {name} </p>
         <p>Address: </p>
         <p>Contact: </p>
-        <p>Tags: </p>
+        <p>Tags: {labels} </p>
       </div>
     )
 }
