+++ src/Components/TherapistPage/TherapistPage.css
@@ -1,23 +1,21 @@
-.provider-box {
-  border: 2px solid indigo;
+.results-window {
+  border: 2px solid sienna;
   width: 80vw;
   display: flex;
   flex-direction: column;
-}
-
-.provider-background {
-  border: 1px solid orange;
-  display: flex;
-  justify-content: space-around;
-}
-
-.provider-information {
-  border: 1px solid blue;
-}
-
-.provider-tags {
-  border: 1px solid red;
-  display: flex;
   align-items: center;
+}
+
+.filter-box {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.results-cards {
+  border: 2px solid salmon;
+  width: 80vw;
+  display: flex;
+  flex-wrap: wrap;
   justify-content: space-evenly;
 }
